+++ src/main/java/de/samply/exporter/ExporterConst.java
@@ -87,6 +87,7 @@ public class ExporterConst {
   public final static String HTTP_RELATIVE_PATH = "HTTP_RELATIVE_PATH";
   public final static String HTTP_SERVLET_REQUEST_SCHEME = "HTTP_SERVLET_REQUEST_SCHEME";
   public final static String TIMESTAMP_FORMAT = "TIMESTAMP_FORMAT";
+  public final static String FHIR_PACKAGES_DIRECTORY = "FHIR_PACKAGES_DIRECTORY";
 
   // Spring Values (SV)
   public final static String HEAD_SV = "${";
@@ -122,6 +123,8 @@ public class ExporterConst {
       HEAD_SV + HTTP_SERVLET_REQUEST_SCHEME + ":http" + BOTTOM_SV;
   public final static String TIMESTAMP_FORMAT_SV =
       HEAD_SV + TIMESTAMP_FORMAT + ":" + DEFAULT_TIMESTAMP_FORMAT + BOTTOM_SV;
+  public final static String FHIR_PACKAGES_DIRECTORY_SV =
+      HEAD_SV + FHIR_PACKAGES_DIRECTORY + ":#{'./fhir-packages'}" + BOTTOM_SV;
 
 
   // REST Paths
@@ -182,6 +185,7 @@ public class ExporterConst {
   public static final String[] PASSWORD_BLACKLIST = {"password", "secret", "apikey"};
   public static final String PASSWORD_REPLACEMENT = "XXXXX";
   public static final String VALIDATION_MESSAGE_SEPARATOR = " | ";
+  public static final String FHIR_PACKAGE_ROOT_CLASSPATH = "fhir-packages";
 
 
 }